--- EquiFood-A-ReactNative/server/routes/Image.js
@@ -1,1 +1,0 @@
-import express, { Router } from "express";
